Time: 28 ms (76.96%) | Memory: 54.2 MB (70.89%) - LeetSync

+++ 1621-number-of-subsequences-that-satisfy-the-given-sum-condition/number-of-subsequences-that-satisfy-the-given-sum-condition.cpp
@@ -11,7 +11,7 @@ class Solution {
 public:
     int numSubseq(vector<int>& nums, int target) {
         int n = nums.size();
-        long long result = 0;
+        int result = 0;
         sort(nums.begin(), nums.end());
 
         vector<int> twoPower = preComputeTwoPower(n);
